act 自定义业务

+++ ruoyi-service/ruoyi-activiti/src/main/java/com/ruoyi/activiti/controller/ActivitiController.java
@@ -1,13 +1,12 @@
 package com.ruoyi.activiti.controller;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import org.activiti.engine.RepositoryService;
 import org.activiti.engine.RuntimeService;
 import org.activiti.engine.repository.ProcessDefinition;
+import org.activiti.engine.repository.ProcessDefinitionQuery;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,10 +14,11 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.ruoyi.activiti.domain.ActReProcdef;
+import com.ruoyi.activiti.service.IActReProcdefService;
 import com.ruoyi.activiti.vo.ReProcdef;
 import com.ruoyi.common.core.controller.BaseController;
 import com.ruoyi.common.core.domain.R;
-import com.ruoyi.common.core.page.PageDomain;
 
 /**
  * 流程控制接口
@@ -31,10 +31,13 @@ import com.ruoyi.common.core.page.PageDomain;
 public class ActivitiController extends BaseController
 {
     @Autowired
-    private RepositoryService repositoryService;
+    private RepositoryService    repositoryService;
 
     @Autowired
-    private RuntimeService    runtimeService;
+    private RuntimeService       runtimeService;
+
+    @Autowired
+    private IActReProcdefService procdefService;
 
     /**
      * 启动一个流程
@@ -50,34 +53,47 @@ public class ActivitiController extends BaseController
         return R.ok();
     }
 
-    @GetMapping("list")
-    public R list(PageDomain page)
+    @GetMapping("allLatest")
+    public R list()
     {
-        List<ProcessDefinition> processDefinitions = repositoryService.createProcessDefinitionQuery()
-                .listPage(page.getPageSize() * (page.getPageNum() - 1), page.getPageSize());
-        long count = repositoryService.createProcessDefinitionQuery().count();
+        ProcessDefinitionQuery query = repositoryService.createProcessDefinitionQuery().latestVersion();
+        List<ProcessDefinition> processDefinitions = query.list();
         List<ReProcdef> list = new ArrayList<>();
         for (ProcessDefinition processDefinition : processDefinitions)
         {
             ReProcdef reProcdef = new ReProcdef(processDefinition);
             list.add(reProcdef);
         }
-        Map<String, Object> m = new HashMap<String, Object>();
-        m.put("rows", list);
-        m.put("pageNum", page.getPageNum());
-        m.put("total", count);
-        return R.ok(m);
+        return R.ok().put("rows", list);
     }
 
-    @PostMapping("delete/{id}")
-    public R deleteOne(@PathVariable("id") String id)
+    @GetMapping("list")
+    public R list(ActReProcdef actReProcdef)
     {
-        // 根据deploymentID删除定义的流程，普通删除
-        repositoryService.deleteDeployment(id);
-        System.out.println("普通删除--流程定义删除成功");
+        startPage();
+        return result(procdefService.selectList(actReProcdef));
+    }
+
+    @PostMapping("remove")
+    public R deleteOne(String ids)
+    {
+        String[] idArr = ids.split(",");
+        for (String id : idArr)
+        {
+            long count = runtimeService.createProcessInstanceQuery().deploymentId(id).count();
+            if (count > 0)
+            {
+                return R.error("流程正在运行中，无法删除");
+            }
+            else
+            {
+                // 根据deploymentID删除定义的流程，普通删除
+                repositoryService.deleteDeployment(id);
+            }
+            // 强制删除
+            // repositoryService.deleteDeployment(id, true);
+            // System.out.println("强制删除--流程定义删除成功");
+        }
         return R.ok();
-        // 强制删除
-        // repositoryService.deleteDeployment(id, true);
-        // System.out.println("强制删除--流程定义删除成功");
     }
 }
